Add more battle phase tests
